+++ utils/js_modules/remove-fallback-warning.mjs
@@ -1,0 +1,7 @@
+const init = () => document.querySelector("body>dialog").remove();
+
+if (window.document.readyState === "loading") {
+  window.document.addEventListener("DOMContentLoaded", init);
+} else {
+  init.apply(window.document);
+}